requirements.txt
discord.py>=2.6.4
yt-dlp>=2023.12.1
Flask>=2.0.0
PyNaCl>=1.5.0
python-dotenv>=1.0.0
